+++ panda_gym/envs/tasks/reach_cam.py
@@ -145,9 +145,9 @@ class ReachCam(Task):
     def compute_reward(self, achieved_goal, desired_goal, info: Dict[str, Any]) -> np.ndarray:
         d = distance(achieved_goal, desired_goal)
         if self.reward_type == "sparse":
-            return -np.array(d > self.distance_threshold, dtype=np.float32)
+            return -np.array(d > self.distance_threshold, dtype=np.float16)
         else:
-            return -d.astype(np.float32)
+            return -d.astype(np.float16)
 
     def get_obj_pos_rotation(self) -> np.ndarray:
         return np.array([])  # no obj related pos or rotation
